mythtv-setup: rename commandlineparser.* to mythtv-setup_…
